[Refactor]Migrated matcher class into Object-Oriented style using Matcher class
Matcher class has been created, all matching processes are now handled by Matcher class. Unit tests have been updated too.
