+++ client/app/src/main/java/com/example/mrides/Activity/CreateRouteActivity.java
@@ -30,6 +30,8 @@ import android.widget.EditText;
 import android.widget.ImageView;
 import android.widget.TextView;
 import android.widget.Toast;
+
+import com.example.mrides.userDomain.Matcher;
 import com.example.mrides.userDomain.PassengerSerializer;
 import com.example.mrides.userDomain.User;
 import com.example.mrides.userDomain.UserSerializer;
@@ -331,8 +333,11 @@ public class CreateRouteActivity extends FragmentActivity implements OnMapReadyC
 
             polylinePaths.add(mGoogleMap.addPolyline(polylineOptions));
 
+            Matcher matcher = new Matcher ();
+            matcher.setMatchedMarkers(this.matchedMarkers);
+            matcher.setUserMapCatalog(this.userOnMapCatalog);
 
-            matchRoute(route.getPoints());
+            matcher.matchRoute(route.getPoints());
         }
     }
 
@@ -412,104 +417,6 @@ public class CreateRouteActivity extends FragmentActivity implements OnMapReadyC
                         "/invitePassenger.php",jsonBody,
                 "application/x-www-form-urlencoded; charset=UTF-8" ,this);
         Toast.makeText(CreateRouteActivity.this, getString(R.string.invie_sent), Toast.LENGTH_SHORT).show();
-    }
-
-    /**
-     * Matches route of driver and passengers
-     * @param routeOfUser
-     */
-    public void matchRoute(List<LatLng> routeOfUser) {
-
-        for (User user : userOnMapCatalog) {
-            ArrayList<Route> passengerRoutes = user.getRoutes();
-
-            for (Route route : passengerRoutes) {
-                LatLng pickUp = route.getStartLocation();
-                LatLng drop = route.getEndLocation();
-                int passengerRouteId = route.getId();
-                boolean pickUpBool = false;
-                boolean goToEnd = false;
-                int i = 0;
-
-                while (i < routeOfUser.size() && pickUpBool == false) {
-                    LatLng pointInPoly = routeOfUser.get(i);
-                    if (validateDistance(pickUp, pointInPoly) && goToEnd == false) {
-                        goToEnd = true;
-                        i++;
-                    }
-
-                    if (validateDistance(drop, pointInPoly) && goToEnd == true) {
-                        for ( int key : matchedMarkers.keySet()) {
-                            if(key == passengerRouteId) {
-                                Marker marker = matchedMarkers.get(key);
-                                marker.setIcon(BitmapDescriptorFactory.defaultMarker(BitmapDescriptorFactory.HUE_ROSE));
-                                pickUpBool = true;
-                                break;
-                            }
-                        }
-                    }
-                    i++;
-                }
-
-            }
-
-        }
-    }
-
-    /** calculates the distance between two locations in MILES */
-    private double distance(double lat1, double lng1, double lat2, double lng2) {
-
-        double earthRadius = 3958.75; // in miles, change to 6371 for kilometer output
-
-        double dLat = Math.toRadians(lat2-lat1);
-        double dLng = Math.toRadians(lng2-lng1);
-
-        double sindLat = Math.sin(dLat / 2);
-        double sindLng = Math.sin(dLng / 2);
-
-        double a = Math.pow(sindLat, 2) + Math.pow(sindLng, 2)
-                * Math.cos(Math.toRadians(lat1)) * Math.cos(Math.toRadians(lat2));
-
-        double c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1-a));
-
-        double dist = earthRadius * c;
-
-        return dist; // output distance, in MILES
-    }
-
-    /**
-     * This method validates the distance between two points, point of a polyline and start and end address
-     * of the passenger. This algorithm indicates that if the distance between two points
-     * satisfy the condition, then this passenger point is part of the polyline
-     * @param passengerLocation
-     * @param userLocation
-     * @return boolean
-     */
-    public boolean validateDistance(LatLng passengerLocation, LatLng userLocation) {
-        if (distance( passengerLocation.latitude, passengerLocation.longitude,
-                userLocation.latitude, userLocation.longitude) <= 0.1) {
-            return true;
-        }
-        else {
-            return false;
-        }
-    }
-
-    /**
-     * A mutator method for UserOnMapCatalog
-     * @param userOnMapCatalog
-     */
-    public void setUserOnMapCatalog (ArrayList <User> userOnMapCatalog){
-        this.userOnMapCatalog = userOnMapCatalog;
-    }
-
-    /**
-     * An accessor method for UserOnMapCatalog
-     * @return userOnMapCatalog
-     */
-
-    public ArrayList <User> getUserOnMapCatalog (){
-        return this.userOnMapCatalog;
     }
 
     /**
